requirements.txt

# Base ------------------------------------------------------------------------
gitpython
ipython  # interactive notebook
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
ultralytics==8.0.106
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.7.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision==0.13
tqdm>=4.64.0
tensorboard>=2.4.1
easyocr==1.5.0

aiofiles==0.7.0
    # via -r requirements.in
asgiref==3.4.1
# Plotting --------------------------------------------------------------------
seaborn>=0.11.0
pydantic==1.8.2
pandas==1.3.3
gunicorn==20.1.0
setuptools>=65.5.1
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.7
    # via requests
click==8.0.1
    # via uvicorn
cycler==0.10.0
    # via matplotlib
fastapi[standard]==0.68.1
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
httptools==0.2.0
    # via uvicorn
idna==3.2
    # via requests
kiwisolver==1.3.2
starlette==0.14.2
six==1.16.0
pytz==2021.3
    # via fastapi
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-dotenv==0.19.0
    # via uvicorn
python-multipart==0.0.5

typing-extensions==3.10.0.2
    # via
    #   pydantic
    #   torch
urllib3==1.26.7
    # via requests
uvicorn[standard]==0.15.0
    # via -r requirements.in
uvloop==0.16.0
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==10.0
    # via uvicorn

PySpice==1.5
